standardize how we lookup environment variables in both frontend and backend
