+++ apps/google-analytics-4/frontend/tsconfig.json
@@ -10,6 +10,6 @@
     "inlineSources": true,
     "sourceRoot": "/src"
   },
-  "include": ["src", "test"],
+  "include": ["src", "test", "../shared/config.ts"],
   "exclude": ["node_modules"]
 }
